# backend/src/main/java/com/securemath/dto/admin/EnrollmentRequestDto.java
package com.securemath.dto.admin;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class EnrollmentRequestDto {
    
    @NotNull(message = "Student ID is required")
    private UUID studentId;
}