# specs/001-secure-math-elearning/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Secure Math Platform API
  version: 1.0.0
  description: Contract for admin and student APIs with strict role segregation.
servers:
  - url: https://api.securemath.local
security:
  - bearerAuth: []
paths:
...
  /api/admin/courses/{courseId}/enrollments:
    post:
      summary: Enroll student in course (Admin only)
      tags: [Admin]
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
      responses:
        '201':
          description: Enrollment created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    EnrollmentRequest:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: string
          format: uuid
...